requirements.txt
alabaster==0.7.7
amqp==1.4.9
ansible==2.1.0.0
anyjson==0.3.3
argcomplete==0.8.7
astroid==1.4.5
autopep8==1.2.1
awscli==1.10.54
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
behave==1.2.5
billiard==3.3.0.23
blinker==1.4
boto==2.36.0
botocore==1.4.44
bypy==1.2.1
celery==3.1.23
celerymon==1.0.3
certifi==14.5.14
chardet==2.3.0
click==3.3
colorama==0.3.7
coverage==4.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.19
django-admin-bootstrapped==2.3.2
django-angular==0.7.15
django-annoying==0.8.0
django-appconf==0.6
django-autocomplete-light==3.1.4
django-bootstrap3==7.0.1
django-celery==3.1.16
django-compressor==1.4
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-devserver==0.8.0
django-el-pagination==2.1.2
django-extensions==1.6.1
django-extra-views==0.6.4
django-filter==0.9.2
django-filters==0.1.0
django-haystack==2.3.1
django-libsass==0.2
django-localflavor==1.1
django-model-utils==2.2
django-mptt==0.6.1
django-oscar==1.1.1
django-pagination==1.0.7
django-pagination-bootstrap==1.0.7
django-pylibmc==0.6.0
django-reversion==1.10.2
django-rq==0.7.0
django-s3-folder-storage==0.2
django-storages==1.1.8
django-tables2==1.0.4
django-tagging==0.3.4
django-tastypie==0.13.3
django-toolbelt==0.0.1
django-treebeard==3.0
django-widget-tweaks==1.3
django-xadmin==0.5.0
djangorestframework==3.3.3
docutils==0.12
easy-thumbnails==2.3
EasyProcess==0.1.6
EbookLib==0.15
ecdsa==0.13
enum34==1.1.6
eventlet==0.19.0
extras==0.0.3
Fabric==1.11.1
factory-boy==2.4.1
feedgenerator==1.8
flake8==2.4.1
flower==0.8.3
futures==2.2.0
Glances==2.6.2
googlemaps==2.1.1
greenlet==0.4.9
gunicorn==19.3.0
guppy==0.1.10
html5lib==0.999
imagesize==0.7.1
influxdb==2.6.0
Jinja2==2.8
jmespath==0.9.0
jsonpickle==0.9.2
kombu==3.0.35
lazy-object-proxy==1.2.2
librabbitmq==1.6.1
libsass==0.6.2
line-profiler==1.0
lxml==3.4.1
Mako==1.0.4
Markdown==2.4.1
MarkupSafe==0.23
mccabe==0.3.1
memory-profiler==0.32
microsofttranslator==0.7
mock==1.0.1
nltk==3.0.5
nodeenv==0.13.6
numpy==1.9.1
oauthlib==0.7.2
paramiko==1.16.0
parse==1.6.6
parse-type==0.3.4
pdfminer==20140328
pdoc==0.3.2
pelican==3.6.3
pep8==1.5.7
phonenumbers==6.3.0
Pillow==2.7.0
ply==3.4
protobuf==3.0.0
psutil==2.2.1
psycopg2==2.8.4
purl==1.1
pyasn1==0.1.9
pyBarcode==0.8b1
pycallgraph==1.0.1
pycrypto==2.6.1
pyflakes==0.8.1
Pygments==2.1.3
pygraphviz==1.2
PyJWT==0.4.1
pylibmc==1.4.1
pylint==1.5.5
Pympler==0.4
PyNLPIR==0.4.4
pysmi==0.0.7
pysnmp==4.3.2
PySocks==1.5.0
python-dateutil==2.5.2
python-docx==0.8.5
python-magic==0.4.12
python-memcached==1.54
python-mimeparse==1.5.1
python-openid==2.2.5
python-pptx==0.5.7
python-social-auth==0.2.2
pytz==2016.4
PyYAML==3.11
redis==2.10.3
reportlab==3.2.0
requests==2.5.1
requests-oauthlib==0.4.2
rq==0.5.1
rsa==3.1.4
s3cmd==1.6.1
s3transfer==0.1.1
scipy==0.16.1
selenium==2.44.0
Shapely==1.5.6
simplejson==3.6.5
six==1.9.0
slimit==0.8.1
smartypants==1.8.6
snowballstemmer==1.2.1
sorl-thumbnail==12.2
SpeechRecognition==1.2.4
Sphinx==1.4.1
sqlparse==0.1.14
static3==0.5.1
statistics==1.0.3.5
statsd==3.1
stem==1.3.0
testtools==1.5.0
textract==1.2.0
tornado==4.1
typogrify==2.0.7
Unidecode==0.4.19
unittest2==0.8.0
urllib3==1.10.2
wrapt==1.10.8
xlrd==0.9.4
XlsxWriter==0.7.2
xlwt==1.0.0
mysqlclient==1.4.6
